requirements.txt
tqdm
pyyaml
torch>=2.0.0
torchvision
tensorboard
